moved decay logic to decay manager
